[FIX] 리세일 판매하기 응답 tokenId로 수정

+++ src/main/java/com/example/demo/domain/resale/dto/response/ResaleInfoDTO.java
@@ -5,6 +5,8 @@ import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
+import java.math.BigInteger;
+
 @Getter
 @Builder
 @NoArgsConstructor
@@ -12,5 +14,5 @@ import lombok.NoArgsConstructor;
 public class ResaleInfoDTO {
 
     private Long resaleId;
-    private Long tokenId;
+    private BigInteger tokenId;
 }
